// src/interfaces/IUser.ts
export interface IUser {
    matricula: string;
    nome: string;
    identidade: string;
    data_exp: string;
    orgao_emissor: string;
    cpf: string;
    sexo: string;
    estado_civil: string;
    data_nasc: string;
    cartao_sus: string;
    endereco: string;
    numero_endereco: string;
    bairro: string;
    cidade: string;
    nome_mae: string;
    nome_pai: string;
    aposentado: string;
    data_cadastro: string;
}